Added pritive Island generation for level one
Added map generation and island generation classes.
Map generation only works for level one as I don't yet know the number
of islands in levels 2 and 3.
The island generator class creates mostly empty classes and we currently
have no models or items made to date so I cannot attach them to the
associated islands. When have more I will be able to continue with this.