Fixed a bug were islands were showing up in the wrong places
Reset the positions for the islands for them to show up appropriately on
the screen.
They are now being generated from script and have the appropriate
backing objects, though they still are not persistant to my knowledge as
they are stored in the MapScreenLoad script.
Maybe should be moved to the gama manager? Or somewhere else that
persists when we navigate away from the map screen.