Functioning "Combat"
If the player has cannons they will show up and pressing the buttons
causes the enemy to take damage.
At the moment the player ship types as well as the enemy ship types have
place holder values which has to change real soon.
As there are no player cannons atm but if you want it to work uncomment
the commented sections in ScreenLoad.SetPlayerCannons and comment out
the foreach loop to get place holder cannons.