The Empire Strikes Back
Enemy ships now have cannons which fire off cooldown at the player.
When the enemy is destroyed they are destroyed with it.
Player ships set to have the stats that the title screen says they do
(save inventory space which doesn't exist atm)
Updated teh ship sink animation to rotate more and sink faster to be
more noticable (though one could turn down the speed if they dont like
it) and to make it seem less like it is simply sinking straight down.